transitioned multiple methods and fixed bugs
- moved a buncha methods from uimanager to gamemanager.
- fixed default save not actually updating when you changed it until you restarted.
- fixed enemies not loading saved number correctly
